Preparing the file
