+++ List of friends/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
