+++ List of friends/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
