+++ List of friends/Resources/Assets.xcassets/Contents.json
