+++ List of friends/Resources/List_of_friends.xcdatamodeld/.xccurrentversion
