+++ List of friends/Resources/List_of_friends.xcdatamodeld/List_of_friends.xcdatamodel/contents
